+++ CHANGELOG.md
@@ -10,8 +10,10 @@
 ```
 - removed stopit from dependencies
 - optimized element actions (now they wait for visibility only if first try failed)
+  - now selene is as fast as selenium with research of elements before any action. Manual cashing is also available when "raw selenium" speed is needed (semi-automatic customizable cashing will be added later)
 - optimized SElementsCollection#find implementation - now it returns first match, not filter everything and then get first among all matches
 - mapped all webelement methods to selement (almost all) - now autocompletion works fully for selement
+- optimized "inner lazy elements" of collection logic, removed a bit magic from it.
 
 ## 0.0.5 (released 20.02.2016)